[29] modified endpoints for jobs API. Wrote an example code snippet for saveJob() (will be updated later).

+++ code/backend/src/main/java/com/cs673/careerforge/request/BaseRequest.java
@@ -10,5 +10,5 @@ import lombok.*;
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 public class BaseRequest {
-    private int uid;
+    private long uid;
 }
